+++ base/src/Extension_Bool.cpp
@@ -27,6 +27,7 @@ limitations under the License.
 #include "Category_Equals.hpp"
 #include "Category_Float.hpp"
 #include "Category_Formatted.hpp"
+#include "Category_Hashed.hpp"
 #include "Category_Int.hpp"
 #include "Category_String.hpp"
 
@@ -66,6 +67,8 @@ ReturnTuple DispatchBool(PrimBool value, const ValueFunction& label,
       return ReturnTuple(Box_Bool(value));
     case CategoryId_Formatted:
       return ReturnTuple(Box_String(value? "true" : "false"));
+    case CategoryId_Hashed:
+      return ReturnTuple(Box_Int(value ? 1000000009ULL : 1000000007ULL));
     default:
       FAIL() << "Bool does not implement " << label;
       __builtin_unreachable();